Split ultralight from IoT Device state
